Switch to kingpin command line parser
Unfortunately this is a breaking change, as this means going from
`-config.file` to `--config.file` syntax, but it needs to happen
eventually. It better aligns the IPMI exporter with the rest of the
Prometheus eco-system.

As a first benefit, expose the command line parameters that control the
logging library.