Add a `bmc-device-id` collector
When the `bmc` collector was extended to attempt to collect the system
firmware revision the resulting IPMI request broke on some older
systems. This collector can be used as a fallback for these systems. For
obvious reasons, the `bmc` and `bmc-device-id` collectors cannot be used
at the same time, which is checked at config load time.

This should finally fix #57.