Implement native collectors with go-ipmi
These are for now considered slightly experimental. They can be enabled
by `--native-ipmi`. The implementation is backed by go-ipmi [1]. For the
time being, this will be a purely optional thing. There are some caveats
(which will be documented), but mostly this should be a fully functional
replacement. It may in some aspects even be superior (or at least become
so in the future), as the exporter has much better control over the IPMI
communication.

[1] https://github.com/bougou/go-ipmi